Fix unit tests following upstream change in QGIS API behavior
A bug fix in QGIS ~3.6 impacted the tests (and performance of R
provider) by forcing format conversion of multi-layer input files
(such as geopackage).

New API has been added upstream to avoid this forced conversion
for providers like R, but it's only available in QGIS 3.10 or later.

Update to use this new API when available, and fix failing unit
tests.